requirements.txt
Django==5.0.3
pymysql==1.1.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
django-shortuuidfield==0.1.3
PyJWT==2.8.0